supplementary/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
defusedxml==0.7.1
et-xmlfile==1.1.0
idna==3.4
lxml==4.9.3
numpy==1.25.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
PyMuPDF==1.23.2
PyMuPDFb==1.23.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.22
pytz==2023.3
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.4
xlrd==2.0.1
XlsxWriter==3.1.2